services/qr-svc/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
qrcode[pil]==7.4.2
Pillow==10.3.0
boto3==1.34.92
python-dotenv==1.0.1
pydantic-settings==2.2.1
